+++ App/Shared/Models/BlockCategory.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+public enum BlockCategory: String, CaseIterable, Codable, Sendable {
+    case hygiene
+    case meal
+    case sport
+    case work
+    case medical
+    case shopping
+    case medication
+    case social
+    case kids
+    case sleep
+    case hydration
+    case housekeeping
+}